[FIX] helpdesk_ux: emails send from stage changes as messages (#130)
Odoo introduce a new change were all the stage's emails are send like
notes. But this is causing is trouble because the respond of the
customer is also taked into account as a note, and not a as message.

In order to avoid this we add some type of patch that return the method
that do this behaviour as it was working before the new feature.