[FIX] account_online_sync_ar: Provider name
When manually update it was always been overtiwrited t "Sync Account" removing the bnnk information that let us to idenfity the provider.

Now we use the id_site in order to update this value, which is good becase it also help us when the account has not been sync we are not calling cataloge/sites info we are using directly a mapping in Odoo to get the bank site info directly.