[ADD] account_online_sync_ar: temporal fix button name
Update Credentials Odoo button in Provider account form view has a bad
spanish translation, we already ask Odoo to fix it but we need to wait.

Until then we have made a temporal change in order to be able to proper
show the translated value of the button to avoid confusion in our
customers.

NOTE: we tried to use i18n_extra but does not work