[FIX] account_online_sync_ar: multi company journals
We have an online provider will multiple journals, each one can be in
a different company.

In this case if the user is not currently logged in one of the companies
in the online journals this one will not be updated when they run a
manual sync.

For the case of the ir,cron is made by super user and does not affect.

Also, resolve a compatibility problem win the the account_journal_security journal
this does not let to the admin user to see and sync transactions of the journals
it can not see

X-original-commit: 0f1e243e04e6d7825c002a7ad2a23e97ead0328e
Part-of: ingadhoc/enterprise-extensions#154