[ADD] account_online_sync_ar: auto_sync new logic
Taking the save scheme as odoo/account_online_synchronization module
we add a new boolean field in the online provider object that let
the user to set/unset in order to activate/deactivate the
automatic synchronization.

* if this boolean is False will not auto update the account with the
cron, the account only will be updated if the user do it manually.

Before this change we have the action_required for the user field,
the problem is that the field is not actually editable by the user
and also is used to store another information: when the user need
to fix something in the sybc obj.

X-original-commit: da0403c202d8b31188332f5533171faf65f1978a
Part-of: ingadhoc/enterprise-extensions#154