[FIX] account_online_sync_ar: Update credentails
Not need to take into account the journal of the related online
accounts. Neither the company were the user is logged.

Only need to take into account the company set in the online provider.
This way the token to conenct to syncfy is is properly generated.

Before this change we were using the journals company and instead
of updating the credential we were creating a new one in another
company paybook user.

X-original-commit: 8699ad7f7197c98f8af9766eaabe5f828122990f
Part-of: ingadhoc/enterprise-extensions#154