[FIX] account_online_sync_ar: deleting in multi company
In the Manage credential wizard, when we were deleting an account
from a different company that the user was logged we was having
a connection error because we were using the paybook user id of
the current company and not the one defined actually in the
wizard sync line.

Now we take into accoun tthe paybook user id already loaded in the
wizard line and is working.

X-original-commit: 38dd5aa3f19ea0d6749676b686b15524d3f89637
Part-of: ingadhoc/enterprise-extensions#154