[ADD] account_online_sync_ar: new max sync date.
Recently we have had a problem where some customers that change the
username that they use to connect to the banks. the accounts and
transactions are the same But the paybook user id is different, also the
ID represents everything.

In order to be able to enable the new credential without missing out on
what we have configured in Odoo, we add a new file paybook_max_date that
let us set a max date that will be used to sync the transactions. This
date is to avoid having duplicated transactions if we try to sync a
early date, also works if manual sync has been made to avoid load does
records twice.

closes ingadhoc/enterprise-extensions#156

X-original-commit: 4239a73354718684777067dce1ff47ca51b64c5b
Signed-off-by: Katherine Zaoral <kz@adhoc.com.ar>