[ADD] account_online_sync_ar: compatiiblity with saltedge
Be able to use both syncfy (paybook) and saltedge service from Odoo.
This is to also be able to manage USA banks.

This PRs go to a modification version of  account_online_synchronization
module, were we remove the hook part to avoid overwrite old models
records and be able to use them together.

We create wiz to let the user to choose the service they want to use

X-original-commit: c9e5aa4e9bcbab1c2824ec8af8c2d11a99a6d2ce
Part-of: ingadhoc/enterprise-extensions#157