[IMP] account_balance_import: remove AR_EDI depend
This module now can be used by any localization, not need to have
the AR localization installed. In order to do this we create a new
method that let us to generic search a res.currency by name, this
method then will be extend by each localization that will needed
in order to add their own rules.

closes ingadhoc/enterprise-extensions#310

X-original-commit: 8685729884d0834d017216f84c1a6856a90402cd
Signed-off-by: pablohmontenegro <pam@adhoc.com.ar>
Signed-off-by: Katherine Zaoral <kz@adhoc.com.ar>